refactor: divided code for maintainability
