+++ data/optimum_distances.txt
@@ -1,0 +1,56 @@
+br17
+39
+
+ft53
+6905
+
+ft70
+38673
+
+ftv33
+1286
+
+ftv35
+1473
+
+ftv38
+1530
+
+ftv44
+1613
+
+ftv47
+1776
+
+ftv55
+1608
+
+ftv64
+1839
+
+ftv70
+1950
+
+ftv170
+2755
+
+kro124
+36230
+
+p43
+5620
+
+rbg323
+1326
+
+rbg358
+1163
+
+rbg403
+2465
+
+rbg443
+2720
+
+ry48p
+14422
