+++ src/interfaces/local_move.py
@@ -1,0 +1,2 @@
+class LocalMove:
+    ...
