+++ src/interfaces/local_optimization.py
@@ -1,0 +1,51 @@
+from collections.abc import Iterable, Hashable
+from typing import TextIO, Optional, Any, Set, List, Tuple
+
+from .local_move import LocalMove
+
+Objective = Any
+
+
+class LocalOptimization:
+    def local_moves(self) -> Iterable[LocalMove]:
+        """
+        Return an iterable (generator, iterator, or iterable object)
+        over all local moves that can be applied to the solution
+        """
+        raise NotImplementedError
+
+    def random_local_move(self) -> Optional[LocalMove]:
+        """
+        Return a random local move that can be applied to the solution.
+
+        Note: repeated calls to this method may return the same
+        local move.
+        """
+        raise NotImplementedError
+
+    def random_local_moves_wor(self) -> Iterable[LocalMove]:
+        """
+        Return an iterable (generator, iterator, or iterable object)
+        over all local moves (in random order) that can be applied to
+        the solution.
+        """
+
+        raise NotImplementedError
+
+    def step(self, lmove: LocalMove) -> None:
+        """
+        Apply a local move to the solution.
+
+        Note: this invalidates any previously generated components and
+        local moves.
+        """
+
+        raise NotImplementedError
+
+    def objective_incr_local(self, lmove: LocalMove) -> Optional[Objective]:
+        """
+        Return the objective value increment resulting from applying a
+        local move. If the objective value is not defined after
+        applying the local move return None.
+        """
+        raise NotImplementedError
