+++ src/solvers.py
